ignore *Test classes that contain no failfast context for now.
Not sure what the best solution is here. I don't want to risk tests not running without it being noticed. But it still should be possible to move a test suite from junit to failfast incrementally without converting all at once.

Maybe the solution is to use a different naming pattern.