always use a fresh dispatcher.
failgood was using the default dispatcher for quite some time but it turns out that this does not work so well in failgood's own test suite, where it calls itself a lot, and always uses the same dispatcher. so for now we are back to always using a fresh dispatcher.